Test for class0 event status
The current class0 tests check for signals when we hit a class 0
condition, which can also be reported as events, when the
SPU_CREATE_EVENTS_ENABLED flags is provided to spu_create.

This change renames the current class0 tests to signify that they're
covering the signal-handling case, and moves the SPE-side code to a
generic class0-apps object. We can then add tests to cover the event
case too.

This also needs a bunch of defines for the event conditions.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>