Add testcase for SPU_CREATE_ISOLATE on non-isolated hardware
Add a test to check for the correct errno when attempting to create an
isolated SPE context on non-isolated hardware.

At present, we're not expecting to run the testsuite on
isolated-enabled hardware, so this test expects the failure in all
cases. We'll have to fix this when we start testing on hardware that
supports isolated mode.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>