New run-tests program
Replace the old run-tests shell script with a C program, which
gives much cleaner output, runs faster, and allows us to do more
with the conditional capabilities.

The capabilites are tested for with cap_* conditionals in the
tests.conf files. At present we have cap_root and cap_affinity.
Capabilties can be negated with ! (eg !cap_root).

Also, add a couple of tests for the new testsuite, and update the
README to match.

Tests are now run in a new working-dir, rather than copying the
entire test tree. PATH is set to the test's base directory, so that
any library programs are available.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>