Turn slowfs into a library
Rather than build a binary that a test will need to execute, turn the
slowfs filesystem into a library, with mount() and unmount() calls.

This means we don't have to worry about paths at runtime. However,
the test Makefile will need to be aware of the necessary libraries at
build-time.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>